Moved menu into the config.
