--- include/menuItem.php
@@ -1,14 +1,0 @@
-<?php
-  class MenuItem
-  {
-    public $url;
-    public $title;
-    public $isCurrent;
-
-    function __construct($url, $title, $isCurrent) {
-      $this->url = $url;
-      $this->title = $title;
-      $this->isCurrent = $isCurrent;
-    }
-  }
-?>
